videomakerauto/requirements.txt
openai>=1.0.0
elevenlabs>=1.0.0
moviepy>=1.0.3
python-dotenv>=1.0.0
